added checks, additional info in UserInfo

+++ ui/SYDE461_UI/CreateNewUser.cs
@@ -27,26 +27,35 @@ namespace SYDE461_UI
         private void button1_Click(object sender, EventArgs e)
         {
             newUser.setUserInfo(textBox1.Text, textBox2.Text);
-            caller.fillUsernameAndPass(this, newUser);
             
             //find better way then exposing storedusers like this
-            caller.storedusers.Add(textBox1.Text, textBox2.Text);
-            
-            //add to user file
-            try
+            if (caller.storedusers.ContainsKey(textBox1.Text))
             {
-                //open file. Login screen should have created if not there so append
-                StreamWriter writestream = new StreamWriter("userlist.txt",true);
-                writestream.WriteLine(textBox1.Text);
-                writestream.WriteLine(textBox2.Text);
-                writestream.Close();
+                MessageBox.Show("This user already exists");
             }
-            catch (Exception ex)
+            else
             {
-                MessageBox.Show(ex.ToString());
-            }
+                caller.storedusers.Add(textBox1.Text, textBox2.Text);
+                caller.fillUsernameAndPass(this, newUser);
+                //add to user file
+                try
+                {
+                    //open file. Login screen should have created if not there so append
+                    StreamWriter writestream = new StreamWriter("userlist.txt", true);
+                    writestream.WriteLine(textBox1.Text);
+                    writestream.WriteLine(textBox2.Text);
+                    writestream.Close();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.ToString());
+                }
+                //Add another exeption handler for if the username is already in the dictionary
 
-            this.Close();
+                this.Close();
+            
+            }
+            
         }
 
         public UserInfo getuserinput()
